requirements.txt
boto3
pillow
pydantic
python-dotenv